+++ CHANGES.md
@@ -20,6 +20,7 @@
 
 - video chat room に simulcast の設定を追加した
 - video chat room に upstream video の latency 関連統計ログを追加した
+- video chat room に `SoraAudioOption` を明示的に記述した
 
 ### CHANGE